20220718work

+++ src/main/webapp/assets/js/imgbox.js
@@ -1,26 +1,58 @@
 let img_files = new Array();
-
 $(function(){
+    $.ajax({
+        url:"/api/user/img/restaurant/0",
+        type:"get",
+        success:function(r) {
+            for(let i = 0; i<r.img_list.length; i++){
+                let tag = 
+                '<div class="aaa111">'
+                +'<div class="restaurant_img" filename="'+r.img_list[i].img_back_name+'" style="background-image:url(/api/img/restaurant/'+r.img_list[i].img_back_name+')">'
+                            +'<button onclick=selectImg("'+r.img_list[i].img_seq+'")>선택</button>'            
+                            +'<button onclick=deleteImg("'+r.img_list[i].img_back_name+'\","'+r.img_list[i].img_seq+'") data-seq='+r.img_list[i].img_seq+'>&times;</button>'
+                            +'</div>'
+                            +'<p>'+r.img_list[i].img_front_name+'</p>'
+                            +'</div>';
+                $(".img_list").append(tag);
+                }
+        },
+        error:function(error) {
+            console.log(error);
+        }
+    })
+
     $("#img_file").change(function(){
         let form = $("#img_form");
         let formData = new FormData(form[0]);
         if($(this).val() == ''||$(this).val() == null) return;
         $.ajax({
-            url:"/images/upload/actor",
+            url:"/api/img/restaurant/upload",
             type:"put",
             data:formData,
             contentType:false,
             processData:false,
-            success:function(result) {
-                if(!result.status) {
-                    alert(result.message);
+            success:function(r) {
+                console.log(r)
+                if(!r.status) {
+                    alert(r.message);
                     return;
                 }
-                let tag = '<div class="actor_img" filename="'+result.file+'" style="background-image:url(/images/actor/'+result.file+')">'
-                +'<button onclick=deleteImg("'+result.file+'")>&times;</button>'
-                +'</div>';
-                img_files.push(result.file);
-                $(".img_list").append(tag);
+                let data ={
+                    img_front_name: r.originName,
+                    img_back_name: r.saveFileName
+                };
+                $.ajax({
+                    url:"/api/user/img/restaurant",
+                    type:"put",
+                    contentType:"application/json",
+                    data:JSON.stringify(data),
+                    success:function(r) {
+                        alert(r.message);
+                        location.reload();
+                    }
+                })
+                
+                img_files.push(r.saveFileName);
             },
             error:function(error) {
                 console.log(error);
@@ -30,23 +62,26 @@ $(function(){
 
 })
 
-function deleteImg(file) {
+function selectImg(file) {
+
+}
+function deleteImg(file,seq) {
+    if(!confirm("이미지를 삭제하시겠습니까?"))return;
+    console.log(seq);
     $.ajax({
-        url:"/images/delete/actor/"+file,
+        url:"/api/img/restaurant/"+file,
         type:"delete",
-        success:function(result) {
-            alert(result.message);
-            if(result.status) {
-                // img파일에서 입력된 파라미터 file값과 동일하지 않은 것들을 걸러내서 새로운 배열을 만들고 img_files에 덮어쓴다
-                img_files = img_files.filter((img)=>(file!=img))
-                $(".img_list").html("");
-                for(let i = 0; i<img_files.length; i++){
-                let tag = '<div class="actor_img" filename="'+img_files[i]+'" style="background-image:url(/images/actor/'+img_files[i]+')">'
-                            +'<button onclick=deleteImg("'+img_files[i]+'")>&times;</button>'
-                            +'</div>';
-                $(".img_list").append(tag);
+        success:function(r) {
+            alert(r.message);
+            console.log(seq);
+            $.ajax({
+                url:"/api/user/img?seq="+seq,
+                type:"delete",
+                success:function(r) {
+                    alert(r.message);
+                    location.reload();
                 }
-            }
+            })
         }
     })
 }
